requirements.txt
pytest
mysql-connector
mysql-connector-python~=8.1.0
Flask
flask-cors
